Add code refactorization
